用户进入游戏房间数据收集

+++ gameServer/db/model/stats/room_log.go
@@ -19,7 +19,7 @@ import (
 type RoomLog struct {
 	RecodeId       int        `db:"recode_id" json:"recode_id"`             // 房间数据记录的Id
 	RoomId         int        `db:"room_id" json:"room_id"`                 // 房间id
-	UserId         int        `db:"user_id" json:"user_id"`                 // 用户索引
+	UserId         int64      `db:"user_id" json:"user_id"`                 // 用户索引
 	RoomName       string     `db:"room_name" json:"room_name"`             //
 	KindId         int        `db:"kind_id" json:"kind_id"`                 // 房间索引
 	ServiceId      int        `db:"service_id" json:"service_id"`           // 游戏标识
